# gamemaster-ai/docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: GameMaster AI API
  version: 0.2.0
  description: Auto-synced API surface generated from app/api route handlers.
servers:
  - url: /api
paths:
...
  /register:
    post:
      summary: POST /api/register
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '429':
          description: Too Many Requests
        '500':
          description: Server Error
...
  /scenarios/mine:
    get:
      summary: GET /api/scenarios/mine
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '429':
          description: Too Many Requests
        '500':
          description: Server Error
...
  /users:
    get:
      summary: GET /api/users
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '429':
          description: Too Many Requests
        '500':
          description: Server Error
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token